Refactoring: BaseRepositoryTest implemented
