+++ app/src/androidTest/java/com/allysonjeronimo/toshop/repository/CategoryRepositoryTest.kt
@@ -1,38 +1,20 @@
 package com.allysonjeronimo.toshop.repository
 
-import androidx.arch.core.executor.testing.InstantTaskExecutorRule
-import androidx.room.Room
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import com.allysonjeronimo.toshop.data.AppDatabase
 import com.allysonjeronimo.toshop.data.entity.Category
 import com.allysonjeronimo.toshop.data.entity.CategoryName
 import junit.framework.Assert.assertEquals
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.asExecutor
-import kotlinx.coroutines.test.TestCoroutineDispatcher
-import kotlinx.coroutines.test.TestCoroutineScope
 import kotlinx.coroutines.test.runBlockingTest
 import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
-import org.junit.runner.RunWith
 
-@ExperimentalCoroutinesApi
-@RunWith(AndroidJUnit4::class)
-class CategoryRepositoryTest {
 
-    @Rule
-    @JvmField
-    var rule = InstantTaskExecutorRule()
+class CategoryRepositoryTest : BaseRepositoryTest(){
+
     private lateinit var categoryRepository: CategoryRepository
-
-    private val testDispatcher = TestCoroutineDispatcher()
-    private val testScope = TestCoroutineScope(testDispatcher)
 
     @Before
     fun setup(){
-        val db = mockDatabase()
+        val db = super.mockDatabase()
         categoryRepository = CategoryDataSource(db.categoryDao())
     }
 
@@ -57,15 +39,6 @@ class CategoryRepositoryTest {
         return categories
     }
 
-    private fun mockDatabase() : AppDatabase{
-        return Room.inMemoryDatabaseBuilder(
-            InstrumentationRegistry.getInstrumentation().targetContext,
-            AppDatabase::class.java)
-            .setTransactionExecutor(testDispatcher.asExecutor())
-            .setQueryExecutor(testDispatcher.asExecutor())
-            .build()
-    }
-
     @Test
     fun insert_categories_with_different_location_names() = testScope.runBlockingTest{
         // Given
@@ -83,7 +56,5 @@ class CategoryRepositoryTest {
         const val CATEGORY_LOCALE_PT_BR = "pt-BR"
         const val CATEGORY_LOCALE_EN_US = "en-US"
     }
-
-
 
 }
